chore(Re-installed Es-lint): Refactored Eslint to get rid of annoying parsing error
