--- client/src/app/components/container/staticBooks.js
@@ -1,30 +1,0 @@
-const BOOKS = [
-
-  {
-    id: 1,
-    title: 'Marvel Secret Empire',
-    category: 'Fiction',
-    author: 'Benny Ogidan',
-    quantity: 20,
-    description: 'Benny is a good boy'
-
-  }, {
-    id: 2,
-    title: 'Marvel Secret Empire 2',
-    category: 'Horror',
-    author: 'Benny Ogidan',
-    quantity: 10,
-    description: 'Benny is a bad boy'
-
-  }, {
-    id: 3,
-    title: 'Marvel Secret Empire',
-    category: 'Fiction',
-    author: 'Benny Ogidan',
-    quantity: 12,
-    description: 'Benny is not a girl'
-
-  }
-];
-
-export default BOOKS;
