+++ client/src/app/components/presentation/common/Book/DisplayBook.js
@@ -20,34 +20,9 @@ export const Book  = (books) =>{
               <p>Title: {books.title}</p>
               <p>Category: {books.category}</p>
               <p>Author: {books.author}</p>
-
             </div>
-
-
           </div>
-
         </div>
-
-
-            // <div>
-
-            //     <Col m={7} s={12} l={5}>
-            //       <Card className ='transparent'
-            //       horizontal header={<CardTitle image={books.image}></CardTitle>}
-            //       actions={[<a href='#'>More Information</a>]}>
-            //           <p>Title: {books.title}</p>
-            //           <p>Category: {books.category}</p>
-            //           <p>Author: {books.author}</p>
-            //           <p >Quantity: {books.quantity}</p>
-            //           <Modal header={`${books.title}`} fixedFooter
-            //           trigger={<Button>Loan</Button>} actions={[<a href='#'>Loan Book &nbsp;</a>]}>
-            //           <p>Description: {books.description}</p>
-            //           </Modal>
-            //       </Card>
-            //     </Col>
-
-
-            // </div>
 
         )
