+++ package.json
@@ -51,6 +51,9 @@
 		"dotenv": "^4.0.0",
 		"enzyme": "^3.1.0",
 		"eslint": "^4.3.0",
+		"eslint-config-airbnb": "^16.1.0",
+		"eslint-plugin-jsx-a11y": "^6.0.2",
+		"eslint-plugin-react": "^7.4.0",
 		"express": "^4.15.3",
 		"express-boom": "^2.0.0",
 		"extract-text-webpack-plugin": "^3.0.0",
@@ -115,8 +118,7 @@
 		"clean-webpack-plugin": "^0.1.16",
 		"coveralls": "^2.13.1",
 		"cross-env": "^5.0.5",
-		"eslint-config-airbnb-base": "^12.0.0",
-		"eslint-plugin-import": "^2.7.0",
+		"eslint-plugin-import": "^2.8.0",
 		"git-flow": "^0.2.0",
 		"mocha-lcov-reporter": "^1.3.0",
 		"nodemailer": "^4.1.0",
